requirements.txt
 Flask
 flask-wtf
 python-dotenv
 RPi.GPIO
 rpi_ws281x
 adafruit-circuitpython-neopixel